[bg-160535014] Improve codebase and achieve a coverage of 100%
